Create README.md
A weather app is a mobile or desktop application that provides users with up-to-date information on the weather conditions in their current location, as well as in other locations they may be interested in. These apps use real-time data from weather stations and satellite imagery to give users accurate and detailed information about temperature, humidity, wind speed, precipitation, and more.